requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
countryinfo==0.1.2
cssselect2==0.8.0
Django==5.0.4
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-phonenumber-field==7.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
fonttools==4.58.2
frozenlist==1.7.0
idna==3.10
loguru==0.7.3
multidict==6.4.4
numpy==2.3.0
pandas==2.3.0
phonenumbers==9.0.7
pillow==11.2.1
propcache==0.3.2
pycountry==24.6.1
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
requests-toolbelt==1.0.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
tinycss2==1.4.0
tinyhtml5==2.0.0
twilio==9.6.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
win32_setctime==1.2.0
yarl==1.20.1
zopfli==0.2.3.post1
